chore: refactor invoice seeder
